- is executed replacement of functions for the greater safety   for MsVC8 and later. Or in other words, has removed   the warning, which gives out MsVC8, branch v2-0-beta

+++ MbsAndWcs.cpp
@@ -31,7 +31,7 @@ size_t _MbsToWcs( wchar_t *wcstr, const char *mbstr, size_t count )
 									  mbstr,
 									  -1,
 									  wcstr,
-									  !wcstr ? 0 : count );
+									  !wcstr ? 0 : (int)count );
 	if ( len > 0 )
 		len--;
 	else if ( wcstr )
